requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-environ==0.4.5
PyJWT==2.8.0
pillow==10.1.0
drf_spectacular==0.27.2

